# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IdeaSoft Case API Documentation",
        "description": "API documentation for IdeaSoft Case Study",
        "contact": {
            "name": "Berk Yağmurlu",
            "email": "berkyagmurlu@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/discounts/calculate/{orderId}": {
            "get": {
                "tags": [
                    "Discounts"
                ],
                "summary": "Calculate order discounts",
                "description": "Calculate applicable discounts for a specific order",
                "operationId": "calculateDiscount",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID of order to calculate discounts for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount calculations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "total_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1999.98
                                                },
                                                "discount_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199.99
                                                },
                                                "final_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1799.99
                                                },
                                                "applied_discounts": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "discount_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Summer Sale"
                                                            },
                                                            "type": {
                                                                "type": "string",
                                                                "example": "total_amount"
                                                            },
                                                            "amount": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 199.99
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}